Edit dart interface Griffon -> Assurance
